Delete block.py , move class Block to indexing.py , as requested and delete df.copy
