+++ asv_bench/benchmarks/indexing.py
@@ -531,4 +531,29 @@ class ChainIndexing:
                 df2["C"] = 1.0
 
 
+from datetime import datetime
+
+
+class Block:
+    params = [
+        (True, "True"),
+        (np.array(True), "np.array(True)"),
+    ]
+
+    def setup(self, true_value):
+        self.df = DataFrame(
+            False,
+            columns=np.arange(500).astype(str),
+            index=date_range("2010-01-01", "2011-01-01"),
+        )
+
+        self.true_value = true_value
+
+    def time_test(self, true_value):
+        """Test time for assigning a slice `True` and `np.array(True)`"""
+        start = datetime(2010, 5, 1)
+        end = datetime(2010, 9, 1)
+        self.df.loc[start:end, :] = true_value
+
+
 from .pandas_vb_common import setup  # noqa: F401 isort:skip
